pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "climate_learn"
version = "1.0.0"
authors = [
    { name = "MINT at UCLA", email = "jason.jewik@ucla.edu" },
    { name = "Tung Nguyen", email = "tungnd@cs.ucla.edu" },
    { name = "Jason Jewik", email = "jason.jewik@ucla.edu" },
    { name = "Hritik Bansal", email = "hbansal@ucla.edu" },    
    { name = "Prakhar Sharma", email = "prakhar6sharma@gmail.com" },
    { name = "Aditya Grover", email = "adityag@cs.ucla.edu" },
]
description = "ClimateLearn: Benchmarking Machine Learning for Weather and Climate Modeling"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [    
    "cdsapi>=0.5.1",
    "dask>=2022.2.0",
    "importlib-metadata==4.13.0",
    "matplotlib>=3.5.3",
    "netcdf4>=1.6.2",
    "pytorch-lightning>=1.9.0",
    "scikit-learn>=1.0.2",
    "timm==0.9.2",
    "tensorboard==2.11.2",
    "wandb>=0.13.9",
    "xarray>=0.20.2",
    "rasterio>=1.3.7",
    "scikit-image",
    "einops",
    "lpips"
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",    
    "flake8>=3.9.2",
    "pytest>=7.2.1",
    "properscoring",
    "build",
    "twine"
]
docs = [
    "ipython>=7.34.0",
    "nbsphinx>=0.8.12",
    "sphinx>=5.3.0",
    "sphinx_rtd_theme>=1.1.1"
]

[project.urls]
"Source" = "https://github.com/aditya-grover/climate-learn"
"Issue Tracker" = "https://github.com/aditya-grover/climate-learn/issues"
"Documentation" = "https://climatelearn.readthedocs.io/en/latest/"
